refactor #27: test 패키지 구조 리팩토링
